Moved MembershipService interface to core package
