+++ src/main/java/cloud/fogbow/ms/core/MembershipService.java
@@ -1,4 +1,4 @@
-package cloud.fogbow.ms;
+package cloud.fogbow.ms.core;
 
 import java.util.List;
